+++ comet_pqc/driver/k2410.py
@@ -125,10 +125,46 @@ class Source(Driver):
         def level(self, value: float):
             self.resource.write(f':SOUR:VOLT:LEV {value:E}')
 
+    class Current(Driver):
+
+        class Range(Driver):
+
+            @property
+            def auto(self) -> bool:
+                return bool(int(self.resource.query(':SOUR:CURR:RANG:AUTO?')))
+
+            @auto.setter
+            @opc_wait
+            def auto(self, value: bool):
+                self.resource.write(f':SOUR:CURR:RANG:AUTO {value:d}')
+
+            @property
+            def level(self) -> float:
+                return float(self.resource.query(':SOUR:CURR:RANG?'))
+
+            @level.setter
+            @opc_wait
+            def level(self, value: float):
+                self.resource.write(f':SOUR:CURR:RANG {value:E}')
+
+        def __init__(self, resource):
+            super().__init__(resource)
+            self.range = self.Range(resource)
+
+        @property
+        def level(self) -> float:
+            return float(self.resource.query(':SOUR:CURR:LEV?'))
+
+        @level.setter
+        @opc_wait
+        def level(self, value: float):
+            self.resource.write(f':SOUR:CURR:LEV {value:E}')
+
     def __init__(self, resource):
         super().__init__(resource)
         self.function = self.Function(resource)
         self.voltage = self.Voltage(resource)
+        self.current = self.Current(resource)
 
     @opc_wait
     def clear(self):