+++ comet_pqc/instruments/instrument.py
@@ -1,0 +1,8 @@
+from abc import ABC
+
+__all__ = ['Instrument']
+
+class Instrument(ABC):
+
+    def __init__(self, context):
+        self.context = context