+++ docs/changelog.md
@@ -7,6 +7,9 @@ nav_order: 10
 # Changelog
 {: .no_toc }
 
+### Unreleased
+- instrument abstraction using abstract class interface.
+
 ## 0.28.0
 
 ### Added